Move the example legacy summary to the diagnostics
`Acceptance` named `souther.compiler.examples` for one thing: the
one-line text a `CompileException` over failing rows carries. That is
not the example subsystem's work. How several failures are gathered into
one sentence is a decision about a report, and it stayed there because
that is where the failures happened to be counted.

Not on `DiagnosticRenderer` either. How one diagnostic becomes a legacy
body is every diagnostic's question; how many of them make "N examples
do not hold" is this one's, and putting it there would teach the general
renderer an aggregation rule about examples.

So `diag.ExampleDiagnostics`, beside `diag.msg.ExampleMessage`, which is
where the vocabulary this renders already lives.

`Acceptance` now names nothing of the example implementation at all, and
`ExampleVerifier` has stopped needing the renderer.

A move. #420 already settled that this text is legacy exception body and
not what a reader is shown; nothing about that changes here.

Refs #971.
